Fix footer components
